[REF] web: convert luxon reactivity tests to HOOT
task-3705027
