+++ addons/web/__manifest__.py
@@ -477,7 +477,6 @@ This module provides the core of the Odoo Web Client.
             'web/static/src/webclient/clickbot/clickbot.js',
         ],
         'web.qunit_suite_tests': [
-            'web/static/tests/legacy/reactivity_tests.js',
             'web/static/tests/legacy/core/**/*.js',
             'web/static/tests/legacy/search/**/*.js',
             ('remove', 'web/static/tests/legacy/search/helpers.js'),
